requirements.txt
appdirs==1.4.3
boto==2.46.1
Django==1.10.6
django-cors-headers==2.0.2
django-extensions==1.7.8
django-ses==0.8.2
djangorestframework==3.6.1
packaging==16.8
plivo==0.11.3
psycopg2==2.7.1
pygraphviz==1.3.1
pyparsing==2.2.0
python-dateutil==2.6.0
requests==2.13.0
six==1.10.0
stripe==1.51.0
